bloqade_lanes_bytecode_core::bytecode

Module opcode

Enums

ArrayInstCode

Array instruction codes.

AtomArrangementInstCode

AtomArrangement instruction codes.

CpuInstCode

Cpu instruction codes (FLAIR-aligned).

DetectorObservableInstCode

DetectorObservable instruction codes.

DeviceCode

Device code identifying which hardware subsystem an instruction targets.

LaneConstInstCode

LaneConstants instruction codes.

MeasurementInstCode

Measurement instruction codes.

OpCodeCategory

Categorized opcode after decoding.

QuantumGateInstCode

QuantumGate instruction codes.

Functions

decode_opcode

Decode a u32 opcode word into an OpCodeCategory.

pack_opcode

Pack a device code and instruction code into a full opcode.